Don't wait for client completePause when sleeping
The client process suspension doesn't wait for WM to be "ready"
(despite the documentation in goToSleepIfPossible). This means
that sometimes the client will NOT report completePause until
the display is turned back on. So we need to assume that if
the display is sleeping, we don't need to wait for client
onPause to consider the activity sync-finished.

Bug: 383245220
Test: with a slow app running, turn screen off and back on quickly
Flag: EXEMPT bugfix
Change-Id: I75ec662ebde04acb7d083643b027ea56e233dbce