[flexiglass] Fix ScrollState#isScrolledToTop on split, and dual shade
ScrollState#isScrolledToTop incorrectly always returned true in split
and dual shade. This was due to a bug where we only considered the
scrim offset value, which is always zero in these configurations.

It was making the NSSL intercept all the downward swipes, which
expanded the notifications instead of scrolling the shade.

Fixes: 380210888
Test: scroll down the shade in split or dual shade
	- have enough notifications in the shade to scroll
        - swipe down when scrolled to the top -> swipe extends notif
        - scroll up a bit, then swipe down -> swipe scrolls the stack
Flag: com.android.systemui.scene_container
Change-Id: Id8e4ba1b6e86d37b9eecc28e0bb041b2e70e139e