[QSDetailedView] Make the current InternetDialogDelegate legacy
This cl did the following thing to prepare for adding the shared
details content manager by the dialog and the non-dialog videw:
1) Added the `assertInNewMode` function in QsDetailedView flag.
2) Renamed `InternetDialogDelegate` -> `InternetDialogDelegateLegacy`.
Added a TODO to `assertInLegacyMode` in the constructor. If we add
this line directly, the UI will crash when the flag is enabled.
3) Corresponding tests change.

Bug:b/377388104
Flag: com.android.systemui.qs_tile_detailed_view
Test: InternetDialogDelegateLegacyTest
Change-Id: I27abec6eb145da42f873d67c490b839e4b8992ab