Update logging in getAnglePackageName()
The ANGLE APK may not be present if ANGLE is the native driver. However,
we still want to attempt to load the ANGLE APK if it's present so we use
the latest/developer libraries. If the APK is not present, this can lead
to misleading log messages such as:

    Invalid number of ANGLE packages. Required: 1, Found: 0

It is not "invalid" for zero ANGLE APKs to be installed when ANGLE is
the native driver.

Update the logging to separate out when:
* No APKs are installed, omitting the "invalid" messaging.
* Too many ANGLE packages are found, since there should never be >1.

This also removes the redundant logging from setupAngleFromApk().

Also cleans up ".get(0)" to ".getFirst()", to remove a warning.

Bug: 372694741
Test: Manual verification of log message with ANGLE as native driver.
Change-Id: Icf820b44c4c7391f08d17d44a59951ba6340a2d0