Remove TestApi annotation from PIC
The TestApi annotation in PropertyInvalidatedCache was added to expose
APIs to a CTS test, in preparation for making the APIs SystemApi.
However, the API review requested that the APIs be encapsulated in a
new class (IpcDataCache), rendering the PIC annotations superfluous.

The following IpcDataCache APIs are now TestApi:
 * forgetDisableLocal()
 * isDisabled()
 * getDisabledState()
 * disableInstance()
 * disableSystemWide()
 * setTestMode()

All TestApi annotations are removed from PropertyInvalidatedCache.
The APIs are visible through IpcDataCache, or are not required by CTS.
(Some APIs are used in unit testing; these are now marked
VisibleForTesting).

Flag: EXEMPT refactor
Bug: 385198070
Test: atest
 * FrameworksCoreTests:PropertyInvalidatedCacheTests
 * FrameworksCoreTests:IpcDataCacheTest
 * CtsOsTestCases:IpcDataCacheTest
Change-Id: I01c88359a784d8bcf788c046eb973cbee0392b24