Fix wrong SysUiState after closing top dialog
When opening two dialogs on top of each other, then closing the top dialog, SYSUI_STATE_DIALOG_SHOWING is set to false, even though it should remain true because there is the first dialog still open.

To fix this, this CL resets the flag to true when a dialog gains window focus.

Bug: 386871258
Test: SystemUIDialogTest
Test: Manual, i.e. verified that back navigation is not disabled after closing top dialog
Flag: EXEMPT bugfix
Change-Id: I8d8c6d77e84564737bf62177ffd5006547169ce1