Log NotificationManager client throttling to statsd
Because only system_server can log counters, we call into system_server to do it. This is itself rate-limited to once a second, so the performance impact should be negligible (and this logging won't make it to nextfood, regardless).

Also switched the logging limit to once per second, so that it matches our other metrics more closely.

Bug: 389918945
Test: atest NotificationManagerTest
Flag: android.app.nm_binder_perf_log_nm_throttling
Change-Id: I2b7961c3fe65fb461e80f7cec31cde1bb07a2e44