Cache notification channel groups.
Similar to the approach for caching channels, this change caches the list of all groups associated with a package. Unlike the getNotificationChannelGroups binder call result, the cached list is not filled in with channel information; we fil that in on the client side for cached data to avoid caching channel data twice.

Bug: 381131846
Flag: android.app.nm_binder_perf_cache_channels
Test: NotificationManagerTest (CTS and unit), PreferencesHelperTest, NotificationManagerServiceTest, manual via looking at dumpsys cacheinfo
Change-Id: I413762351f0a862f4073ce0f8bbf74b9d8246fa8