[SB] Remove StatusBarNoHunBehavior's dependency on app icons in notifs.
StatusBarNoHunBehavior is dependent on notificationsRedesignAppIcons from a product perspective (if we don't do app icons in notifications, then we shouldn't remove the HUN behavior) but eng-wise it's fine for them to be on separately. And this means we don't have to enable both flags for every single unit test (see fixed bug).

Fixes: 390196278
Bug: 385740230
Flag: com.android.systemui.status_bar_no_hun_behavior

Test: atest HomeStatusBarViewModelTest HeadsUpAppearanceControllerTest
CollapsedStatusBarFragmentTest

Change-Id: I10f54bd401c02dc8ac6950888dd898183e936dfd