Close and undo guts on configuration changed
If a guts view like snooze options or notification info was exposed when
the theme or the font changed, it would result in a seemingly empty
notification.

Although we had an onDensityOrFontScaleChanged that was supposed to at
least re-bind the guts when the font was changed, this actually did
nothing because when the guts were re-inflated, the "exposed" state of
the guts was getting reset (so this method was never called).

Now we just undo the changes and close the guts, to make sure they look
correct the next time they are opened.

Bug: 379267630
Test: snooze notification and change the theme via QS
Test: long press notification and change theme/font via adb
Test: atest NotificationSnoozeTest NotificationGutsManagerTest
Flag: com.android.systemui.notification_undo_guts_on_config_changed
Change-Id: I7e3caaf8895a3756bb9f4ad4f627cd66d44a8748