Fix IME visibility issue when device is locked
When the device is locked, if a bubble is expanded and the IME is
visible we now hide the IME before collapsing.

Previously we updated Bubbles internal state, then hid the IME and
once the IME visibility state changed we collapsed the bubble.
But the IME state gets frozen when the device is locked, so that
resulted in corrupt IME and Bubbles state and when the device is
unlocked, the IME remains visible and the bubble remains expanded.

To work around the frozen IME state, this change hides the IME
directly through DisplayImeController to make sure the state
updates correctly.

Flag: EXEMPT bug fix
Fix: 391062949
Test: manual
       - have a floating bubble
       - expand and show IME
       - lock device
       - unlock and observe bubble is collapsed
Test: BubbleControllerTest is being created in a separate change
       - will add unit tests once it's submitted
       - tracked in b/392605916
Change-Id: I955996c8b794832ac5429f06ddb13450f1371dc4