Register For JankData
This change registers a listener that will receive batches of JankData
from Surfaceflinger. The registratoin and unregistration of the callback
is tied to the lifecycle of the enclosing activity. When an activity is
created or resumed a new listener will be registered. When an activity
is pasued or deleted the callback will be unregistered.

When a batch of JankData is received it is passed to the
JankDataProcessor, which will perform frame metric aggregation.

Bug: 376115668
Flag: android.app.jank.detailed_app_jank_metrics_logging_enabled
Test: atest CoreAppJankTestCases
Change-Id: I76ea652bdf54abe3e03d76c5125d69b029babd34