Trigger PowerManager setting update from DreamManager.
Currently, PowerManager relies on a broadcast receiver to be
notified when settings change. In a testing environment, this
can lead to inconsistent behavior as there is no guarantees
around the timing for receiving the broadcast.

This changelist informs PowerManager of a settings change when
screensaver enabled settings are changed from DreamManager.

Test: atest DreamServiceTest
Bug: 393270901
Flag: EXEMPT bugfix
Change-Id: I83cf24dc8fc70aab3e183ba9821b1c8162b6db8b