Specify event for latch in InputMethodServiceTest
Currently we set and wait on a single countDownLatch for either of
onStartInputView, onFinishInputView or onConfigurationChanged. This can
introduce some noise if a different event than the expected one is
trigerred (e.g. onConfigurationChanged while waiting for
onStartInputView). This also introduces some ambiguity as it's not clear
which event is the expected one for each test.

This introduces an IntDef to specify the expected event, and only
triggeres the latch for that specific event.

Flag: EXEMPT testfix
Bug: 394328311
Test: atest InputMethodServiceTest
Change-Id: If435214366422943420a68e14ad712ce4d97fb0d