Send topology to newly registered listener
This commit adds a mechanism to send the current `DisplayTopology` to
newly registered `OnDisplaysChangedListener` instances.

When a listener is added via `addDisplayWindowListener`, it now
receives a call to `onTopologyChanged` with the current
`DisplayTopology`. This ensures that listeners are immediately aware
of the display topology upon registration, without having to wait for
a topology change event.

Bug: 383069173
Test: manual and atest
Flag: com.android.window.flags.enable_connected_displays_window_drag
Change-Id: I3c578de2786c17257cf6cec0e30efdf7832efa5f